Extract parsing function into container-level helper
